+++ common/src/main/java/traben/entity_model_features/config/EMFConfig.java
@@ -135,7 +135,7 @@ public class EMFConfig extends TConfig {
                                 new TConfigEntryBoolean("entity_model_features.config.only_client", "entity_model_features.config.only_client.tooltip",
                                         () -> onlyClientPlayerModel, value -> onlyClientPlayerModel = value, false),
                                 new TConfigEntryBoolean("entity_model_features.config.reset_player", "entity_model_features.config.reset_player.tooltip",
-                                        () -> resetPlayerModelEachRender, value -> resetPlayerModelEachRender = value, false)
+                                        () -> resetPlayerModelEachRender, value -> resetPlayerModelEachRender = value, true)
                         ),
                         new TConfigEntryCategory("entity_model_features.config.performance").add(
                                 new TConfigEntryEnumSlider<>("entity_model_features.config.update", "entity_model_features.config.update.tooltip",